requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
plotly>=5.15.0

# Desktop GUI
PySide6>=6.5.0
pyqtgraph>=0.13.0

# Geospatial analysis
geopandas>=0.13.0
shapely>=2.0.0
mapboxgl>=0.10.0

# Data processing
dask[dataframe]>=2023.12.0
joblib>=1.3.0
openpyxl>=3.1.0

# PDF processing
PyMuPDF>=1.23.0
reportlab>=4.0.0
charset-normalizer>=3.3.0

# Machine Learning
xgboost>=2.0.0
shap>=0.43.0
faiss-cpu>=1.7.0

# 3D Visualization
open3d>=0.17.0
trimesh>=4.0.0
pyvista>=0.43.0

# Geospatial
geopandas>=0.14.0
folium>=0.15.0

# Optimization
scipy>=1.11.0
deap>=1.4.0

# Web3 and utilities
web3>=6.15.0

# PyTorch (latest stable)
torch>=2.1.0
torchvision>=0.16.0

# Hugging Face and LLM
transformers>=4.35.0
huggingface_hub>=0.17.0
python-dotenv>=1.0.0
accelerate>=0.24.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
bitsandbytes>=0.41.0
sympy>=1.12.0
tavily-python>=0.3.0
duckduckgo-search>=5.0.0

# Additional utilities
requests>=2.31.0
tqdm>=4.66.0
Pillow>=10.0.0

# Document generation and formatting
python-docx>=1.1.0
reportlab>=4.0.0
fpdf>=1.7.0
tabulate>=0.9.0
rich>=13.7.0
prettytable>=3.9.0
markdown>=3.5.0

# Build tools
pyinstaller>=6.0.0
cx_Freeze>=7.0.0</content>
<parameter name="filePath">C:\Users\Admin\Desktop\logiciel\requirements.txt